example: switch to modernc.org/sqlite
The modernc.org/sqlite driver was a drop-in replacement
that removes a CGO dependency. Initially, when trying to
follow the README instructions in the example module, I
was seeing the following error:

$ go run .
go: downloading github.com/mattn/go-sqlite3 v1.14.24
2025/11/15 18:40:29 Initializing database: creating schema: Binary was compiled with 'CGO_ENABLED=0', go-sqlite3 requires cgo to work. This is a stub
exit status 1

This is unfortunate not only because CGO in general causes
problems, but because it's a runtime error. Replacing the
sqlite driver removed the error without any other changes.